4 85 34 27 28 29 30 0.1.2 Add test of rm_tx_local_push
Add test suite 8, test [2] (test [1] TODO)

+++ test/include/test_rm7.h
@@ -31,7 +31,7 @@
 #define RM_TEST_FNAMES_N            15
 const char* rm_test_fnames[RM_TEST_FNAMES_N];
 uint32_t    rm_test_fsizes[RM_TEST_FNAMES_N];
-uint32_t    rm_test_L_blocks[RM_TEST_L_BLOCKS_SIZE];
+size_t      rm_test_L_blocks[RM_TEST_L_BLOCKS_SIZE];
 
 struct test_rm_file {
     FILE    *f;
@@ -40,9 +40,9 @@ struct test_rm_file {
 
 struct test_rm_state
 {
-	uint32_t	        *l;
-	void                *buf;
-	void                *buf2;
+    size_t              *l;
+    void                *buf;
+    void                *buf2;
     struct rm_session   *s;
     struct test_rm_file f_z;
 };
